add navbar. allow table to render and sort with real data
