+++ client/src/components/Admin/AdminLanding/AdminLanding.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import {Route, Link} from "react-router-dom";
 import AdminSchedule from "../AdminSchedule/AdminSchedule.jsx";
 import AdminJob from "../../Job/AdminJob.jsx";
+import JobsTable from "../../Shared/JobsTable.jsx";
 
 class AdminLanding extends React.Component {
   constructor(props) {
@@ -16,6 +17,7 @@ class AdminLanding extends React.Component {
           <li><Link to={`${this.props.match.url}/jobDetails`}>JobDetail</Link></li>
         </ul>  
         <Route path={`${this.props.match.url}/:adminSubPages`} component={AdminJob} />
+        <JobsTable/>
       </div>
     );
   }
