--- client/src/components/Admin/AdminLanding/AdminTodayTableEntry.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-class AdminLandingTodayTableEntry extends React.Component {
-  render() {
-    return (
-      <div>
-        FILL_ME_IN
-      </div>
-    );
-  }
-}
-
-export default AdminLandingTodayTableEntry;
