+++ client/src/components/App.jsx
@@ -6,7 +6,9 @@ import SchoolLanding from "./School/SchoolLanding/SchoolLanding.jsx";
 import SubLanding from "./Sub/SubLanding/SubLanding.jsx";
 import AdminLanding from "./Admin/AdminLanding/AdminLanding.jsx"
 import PrivateRoute from "./PrivateRoute.jsx";
-import AdminSubsSummary from "./Admin/AdminSubsSummary/AdminSubsSummary.jsx"
+import AdminSubsSummary from "./Admin/AdminSubsSummary/AdminSubsSummary.jsx";
+import NavBar from "./Menu/NavBar.jsx";
+
 
 class App extends React.Component {
   constructor(props) {
