+++ client/src/components/Home/HomeLanding.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Route, Link} from "react-router-dom";
+import NavBar from "../Menu/NavBar.jsx";
 
 class HomeLanding extends React.Component {
   constructor(props) {
@@ -8,6 +9,7 @@ class HomeLanding extends React.Component {
   render() {
     return (  
       <div>
+        <NavBar/>
         <ul>
           <li><Link to="/admin">Admin</Link></li>
           <li><Link to="/school">School</Link></li>
