+++ client/src/styles/NavBarStyle.js
@@ -1,0 +1,23 @@
+export const styles = theme => ({
+  root: {
+    flexGrow: 1,
+  },
+  flex: {
+    flex: 1,
+  },
+  menuButton: {
+    marginLeft: -12,
+    marginRight: 20,
+  },
+  list: {
+    width: 250,
+  },
+  fullList: {
+    width: 'auto',
+  },
+  sideBar:{
+    width: 250,
+    maxWidth: 360,
+    backgroundColor: theme.palette.background.paper,
+  }
+});
