+++ client/src/styles/ToolbarStyle.js
@@ -1,0 +1,44 @@
+import { lighten } from '@material-ui/core/styles/colorManipulator';
+
+export const toolbarStyle = theme => ({
+  root: {
+    paddingRight: theme.spacing.unit,
+  },
+  highlight:
+    theme.palette.type === 'light'
+      ? {
+          color: theme.palette.secondary.main,
+          backgroundColor: lighten(theme.palette.secondary.light, 0.85),
+        }
+      : {
+          color: theme.palette.text.primary,
+          backgroundColor: theme.palette.secondary.dark,
+        },
+  spacer: {
+    flex: '1 1 100%',
+  },
+  actions: {
+    color: theme.palette.text.secondary,
+  },
+  title: {
+    flex: '0 0 auto',
+  },
+});
+
+export const tableStyle = theme => ({
+  root: {
+    width: '88%',
+    marginTop: theme.spacing.unit * 3,
+  },
+  table: {
+    width: '100%'
+  },
+  tableWrapper: {
+    overflowX: 'auto',
+  },
+  row: {
+    '&:nth-of-type(odd)': {
+      backgroundColor: theme.palette.background.default,
+    },
+  },
+});
